moved contants to another file
